update script: can extract files + run setup

+++ Charlie 2 - Microsoft exchange/Scripts/Installatiescript_JDM.ps1
@@ -1,9 +1,16 @@
+#Check if exchangefiles already exist
 if(Test-Path ExchangeFiles)
 {
-Remove-Item ExchangeFiles 
+$directoryInfo = Get-ChildItem ExchangeFiles | Measure-Object
+if($directoryInfo.count -eq 0)
+{
+	Remove-Item ExchangeFiles 
 }
-ECHO "Checking pre-requisites"
+}
 
+
+#Checking pre-requisites
+ECHO "Checking pre-requisites"
 ECHO "Installing OS roles and features"
 Install-WindowsFeature AS-HTTP-Activation, Desktop-Experience, 
  RPC-over-HTTP-proxy, Web-Mgmt-Console, WAS-Process-Model, Web-Asp-Net45,
@@ -32,33 +39,63 @@ ECHO "RSAT-ADDS feature for AD"
 Install-WindowsFeature RSAT-ADDS, RSAT-Clustering, RSAT-Clustering-CmdInterface, RSAT-Clustering-Mgmt,
 RSAT-Clustering-PowerShell
 ECHO "done installing RSAT-ADDS"
-
 ECHO "DONE checking Pre-Requisites"
 
-new-item ExchangeFiles -itemtype directory 
-ECHO "Place Setupfiles in C:\Users\Administrator\Documents\ExchangeFiles"
 
-<#
-#setup /PrepareAD /OrganizationName:"Red" /IAcceptExchangeServerLicenseTerms
+#making exchangefiles folder
+ECHO "Making folder for setupfiles"
+if(!(Test-Path ExchangeFiles))
+{
+new-item ExchangeFiles -itemtype directory 
+}else
+{
+ECHO "Already created"
+}
+ECHO "Place Setupfiles in C:\Users\Administrator\Documents\ExchangeFiles "
 
 Write-Host "Press any key to continue"
-
 $x = $host.UI.RawUI.ReadKey("NoEcho,IncludeKeyDown")
-
 Write-Host
-Write-Host "A"
-Write-Host "B"
-Write-Host "C"
 
-#>
+#running setupfiles
+start-sleep -m 5
+.\ExchangeFiles\Exchange2016-x64.exe 
+
+#entering DIR & pressing enter
+start-sleep -m 50000
+$wshShell = new-object -com wscript.shell
+$wshShell.SendKeys("C:\Users\Administrator\Documents\ExchangeFiles")
+start-sleep -m 5
+$wshShell.SendKeys("{Enter}") 
+$wshShell.SendKeys("{Enter}")
+ECHO "Extracting"
+
+######NOT DONE########
+#Checking if extraction is complete
+start-sleep -m 500
+$ProcessActive = Get-Process |  Where {$_.name -Match "Extracting"} | Format-Wide -Column 1
+$CheckNull = $ProcessActive -eq $false
+
+do
+{
+ECHO "test"
+start-sleep -m 50000
+$ProcessActive = Get-Process |  Where {$_.name -Match "Extracting"} 
+$CheckTrue = $ProcessActive -eq $True
+ECHO "Still extracting"
+}
+while($CheckNull -eq $true)
+####################
+
+#Running setup.exe
+.\ExchangeFiles\setup /PrepareAD /OrganizationName:"Red" /IAcceptExchangeServerLicenseTerms
 
 
+
+
+#Restarting PC !NEED TO BE AT END!
 Write-Host "Press any key to restart"
-
 $x = $host.UI.RawUI.ReadKey("NoEcho,IncludeKeyDown")
-
 Write-Host
-Write-Host "A"
-Write-Host "B"
-Write-Host "C"
+
 Restart-Computer
